code/src/app/services/rule_validator.py
import pandas as pd
import openai
from typing import List, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

class RuleValidator:
    def __init__(self):
        openai.api_key = os.getenv("OPENAI_API_KEY")
        
    def interpret_rules(self, rules_df: pd.DataFrame) -> List[Dict[str, Any]]:
        """Interpret rules using OpenAI API"""
        rules_text = rules_df.to_string()
        prompt = f"""
        Analyze these data validation rules and convert them into a structured format:
        {rules_text}
        
        For each rule, provide:
        1. The condition to check
        2. The columns involved
        3. The expected format or values
        """
        
        try:
            response = openai.ChatCompletion.create(
                model="gpt-3.5-turbo",
                messages=[
                    {"role": "system", "content": "You are a data validation expert."},
                    {"role": "user", "content": prompt}
                ]
            )
            
            # Parse the response and structure it
            interpreted_rules = self._parse_rules_response(response.choices[0].message.content)
            return interpreted_rules
            
        except Exception as e:
            print(f"Error interpreting rules: {str(e)}")
            return []
    
    def validate_data(self, df: pd.DataFrame, rules: List[Dict[str, Any]]) -> pd.DataFrame:
        """Apply rules to the dataset and flag violations"""
        result_df = df.copy()
        
        # Add columns for rule validation
        result_df['rule_violation'] = 0
        result_df['violation_justification'] = ''
        
        for rule in rules:
            # Apply each rule and update flags
            mask = self._apply_rule(df, rule)
            result_df.loc[mask, 'rule_violation'] = 1
            result_df.loc[mask, 'violation_justification'] += f"{rule['description']}; "
        
        return result_df
    
    def _apply_rule(self, df: pd.DataFrame, rule: Dict[str, Any]) -> pd.Series:
        """Apply a single rule to the dataset"""
        # This is a placeholder - implement specific rule application logic
        # based on the rule type and conditions
        return pd.Series(False, index=df.index)
    
    def _parse_rules_response(self, response: str) -> List[Dict[str, Any]]:
        """Parse the OpenAI response into structured rules"""
        # This is a placeholder - implement proper parsing logic
        return [] 