code/src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.2
plotly==5.18.0
uvicorn==0.34.0
isolation-forest==0.1.0 